+++ semantics-kafka/src/main/scala/atiesh/source/KafkaSource.scala
@@ -27,5 +27,4 @@ class KafkaSource(
   def startup(): Unit = { 
     logger.info("starting source <{}>", getName)
   }
-
 }